% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecovisio_import.R
\name{ecovisio_import}
\alias{ecovisio_import}
\title{Import EcoVisio CSV file}
\usage{
ecovisio_import(path)
}
\arguments{
\item{path}{the name of the file which the data are to be read from}
}
\description{
\code{ecovisio_import} imports a CSV data file exported from EcoVisio 4 and
transforms the data into long format
}
\examples{
EcoVisioImport()
}
