% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoVisioImport.R
\name{EcoVisioImport}
\alias{EcoVisioImport}
\title{Import EcoVisio CSV file}
\usage{
EcoVisioImport(file)
}
\arguments{
\item{file}{the name of the file which the data are to be read from}
}
\description{
\code{EcoVisioImport}Imports an EcoVisio 4 CSV export data file and
transforms into long format
}
\examples{
EcoVisioImport()
}
